Add Ability to View Other Critter Types
Add the ability to view other critter types in the webapp. Achieve this by adding icons to the critter page that represent the different critter types. When an icon is clicked, switch the displayed critters to the icon's type. Make the critter cards only show table rows applicable to that critter type. There should be no empty table cells.