Add Critter Filtering and Donating
Add automatic critter filtering so that the webapp only displays critters that are available today and available right now. For the moment, this property cannot be toggled.

Add the ability to toggle between displaying Northern and Southern month ranges for critters. This is toggled by a checkbox.

Add the ability to toggle filtering for donated critters. By default, donated critters will be filtered out from the displayed critters. When a critter card donated button is pressed, the card will disappear. This filter can be toggled off to show donated critters with a checkbox at the top of the webpage.

Add an informative message to the top of the page that helps the user understand which critters are being shown and what filters are being applied.